tp3/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
idna==2.10
mysql-connector-python==8.0.22
protobuf==3.14.0
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.26.2
